feat: use double for bigger numbers

+++ character-count.c
@@ -3,9 +3,9 @@
 /* count all characters in input before EOF */
 int main ()
 {
-	long nc;
+	double nc;
 
 	for (nc = 0; getchar () != EOF; ++nc)
 		;
-	printf ("%ld\n", nc);
+	printf ("%.0f\n", nc);
 }
